requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
apipkg==1.5
    # via
    #   eth-brownie
    #   execnet
appdirs==1.4.4
    # via
    #   black
    #   eth-brownie
astroid==2.5.2
    # via pylint
asttokens==2.0.4
    # via
    #   eth-brownie
    #   vyper
attrs==20.3.0
    # via
    #   black
    #   eth-brownie
    #   hypothesis
    #   jsonschema
    #   pytest
autopep8==1.5.6
    # via -r requirements.in
bandit==1.7.0
    # via -r requirements.in
base58==2.1.0
    # via
    #   eth-brownie
    #   multiaddr
bitarray==1.2.2
    # via
    #   eth-account
    #   eth-brownie
black==19.10b0
    # via eth-brownie
certifi==2020.12.5
    # via
    #   eth-brownie
    #   requests
chardet==4.0.0
    # via
    #   eth-brownie
    #   requests
click-log==0.3.2
    # via -r requirements.in
click-option-group==0.5.2
    # via -r requirements.in
click-plugins==1.1.1
    # via -r requirements.in
click-spinner==0.1.10
    # via -r requirements.in
click==7.1.2
    # via
    #   black
    #   click-log
    #   click-option-group
    #   click-plugins
    #   eth-brownie
cytoolz==0.11.0
    # via
    #   eth-brownie
    #   eth-keyfile
    #   eth-utils
eth-abi==2.1.1
    # via
    #   -r requirements.in
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   web3
eth-account==0.5.2
    # via
    #   eth-brownie
    #   web3
eth-brownie==1.14.3
    # via -r requirements.in
eth-event==1.2.1
    # via eth-brownie
eth-hash[pycryptodome]==0.2.0
    # via
    #   eth-brownie
    #   eth-event
    #   eth-utils
    #   web3
eth-keyfile==0.5.1
    # via
    #   eth-account
    #   eth-brownie
eth-keys==0.3.3
    # via
    #   eth-account
    #   eth-brownie
    #   eth-keyfile
eth-rlp==0.2.1
    # via
    #   eth-account
    #   eth-brownie
eth-typing==2.2.2
    # via
    #   eth-abi
    #   eth-brownie
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==1.9.5
    # via
    #   eth-abi
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
execnet==1.8.0
    # via
    #   eth-brownie
    #   pytest-xdist
gitdb==4.0.7
    # via gitpython
gitpython==3.1.14
    # via bandit
hexbytes==0.2.1
    # via
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-rlp
    #   web3
hypothesis==5.41.3
    # via eth-brownie
idna==2.10
    # via
    #   eth-brownie
    #   requests
inflection==0.5.0
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
iniconfig==1.1.1
    # via
    #   eth-brownie
    #   pytest
ipfshttpclient==0.7.0
    # via
    #   eth-brownie
    #   web3
isort==5.8.0
    # via pylint
jsonschema==3.2.0
    # via
    #   eth-brownie
    #   mythx-models
    #   web3
lazy-load==0.8.2
    # via -r requirements.in
lazy-object-proxy==1.6.0
    # via
    #   astroid
    #   lazy-load
lru-dict==1.1.7
    # via
    #   eth-brownie
    #   web3
mccabe==0.6.1
    # via pylint
more-itertools==8.7.0
    # via
    #   eth-brownie
    #   pytest
multiaddr==0.0.9
    # via
    #   eth-brownie
    #   ipfshttpclient
mythx-models==1.9.1
    # via
    #   eth-brownie
    #   pythx
netaddr==0.8.0
    # via
    #   eth-brownie
    #   multiaddr
packaging==20.9
    # via
    #   eth-brownie
    #   pytest
parsimonious==0.8.1
    # via
    #   eth-abi
    #   eth-brownie
pathspec==0.8.1
    # via
    #   black
    #   eth-brownie
pbr==5.5.1
    # via stevedore
pluggy==0.13.1
    # via
    #   eth-brownie
    #   pytest
prompt-toolkit==3.0.8
    # via eth-brownie
protobuf==3.15.6
    # via
    #   eth-brownie
    #   web3
psutil==5.8.0
    # via eth-brownie
py-solc-ast==1.2.8
    # via eth-brownie
py-solc-x==1.1.0
    # via eth-brownie
py==1.10.0
    # via
    #   eth-brownie
    #   pytest
    #   pytest-forked
pycodestyle==2.7.0
    # via autopep8
pycryptodome==3.10.1
    # via
    #   eth-brownie
    #   eth-hash
    #   eth-keyfile
    #   vyper
pygments-lexer-solidity==0.5.1
    # via eth-brownie
pygments==2.6.1
    # via
    #   eth-brownie
    #   pygments-lexer-solidity
pyjwt==1.7.1
    # via
    #   eth-brownie
    #   pythx
pylint==2.7.3
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   eth-brownie
    #   packaging
pyrsistent==0.17.3
    # via
    #   eth-brownie
    #   jsonschema
pytest-forked==1.3.0
    # via
    #   eth-brownie
    #   pytest-xdist
pytest-timeout==1.4.2
    # via -r requirements.in
pytest-xdist==1.34.0
    # via eth-brownie
pytest==6.0.1
    # via
    #   eth-brownie
    #   pytest-forked
    #   pytest-timeout
    #   pytest-xdist
python-dateutil==2.8.1
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
python-dotenv==0.16.0
    # via -r requirements.in
pythx==1.6.1
    # via eth-brownie
pyyaml==5.4.1
    # via
    #   bandit
    #   eth-brownie
regex==2020.11.13
    # via
    #   black
    #   eth-brownie
requests==2.25.1
    # via
    #   eth-brownie
    #   ipfshttpclient
    #   py-solc-x
    #   pythx
    #   vvm
    #   web3
rlp==1.2.0
    # via
    #   eth-account
    #   eth-brownie
    #   eth-rlp
semantic-version==2.8.5
    # via
    #   eth-brownie
    #   py-solc-x
    #   tokenlists
    #   vvm
    #   vyper
six==1.15.0
    # via
    #   asttokens
    #   bandit
    #   eth-brownie
    #   jsonschema
    #   multiaddr
    #   parsimonious
    #   protobuf
    #   pytest-xdist
    #   python-dateutil
smmap==4.0.0
    # via gitdb
sortedcontainers==2.3.0
    # via
    #   eth-brownie
    #   hypothesis
stevedore==3.3.0
    # via bandit
tokenlists==0.1.0a1
    # via -r requirements.in
toml==0.10.2
    # via
    #   autopep8
    #   black
    #   eth-brownie
    #   pylint
    #   pytest
toolz==0.11.1
    # via
    #   cytoolz
    #   eth-brownie
tqdm==4.53.0
    # via eth-brownie
typed-ast==1.4.2
    # via
    #   black
    #   eth-brownie
urllib3==1.26.4
    # via
    #   eth-brownie
    #   requests
varint==1.0.2
    # via
    #   eth-brownie
    #   multiaddr
vvm==0.1.0
    # via eth-brownie
vyper==0.2.11
    # via eth-brownie
wcwidth==0.2.5
    # via
    #   eth-brownie
    #   prompt-toolkit
web3==5.11.1
    # via eth-brownie
websockets==8.1
    # via
    #   eth-brownie
    #   web3
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
